Create #14
Write a program to find the smallest positive integer that satisfies the following property: If
you take the leftmost digit and move it all the way to the right, the number thus obtained is
exactly 3.5 times larger than the original number. For instance, if we start with 2958 and move
the 2 all the way to the right, we get 9582, which is roughly 3.2 times the original number.